+++ tsconfig.json
@@ -21,7 +21,8 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./src/*"],
-      "@/backend*": ["./backend/*"]
+      // "@/backend*": ["src/backend/*"]
+      "@/backend": ["./src/backend/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
